Update MessageReplaySubscriber.c
I added more comments and the test setup and test itself. Also in the main.  Made the comment style consistent (all C not C++).  

If this is derived from QueuePublisher and QueueSubscriber then those samples need work too. For starter remove any reference to the deprecated QNN.

Other general style things,  pointers should be initialized to NULL not 0,  subCode should be initialized and compared to member of the enum (SOLCLIENT_SUBCODE_OK) not zero.